Edited Pref Dialog scope to persist past SideMenu
